Deleted old l2 chain. Renamed references.
